# website/public/openapi.json
{
  "components": {
    "schemas": {
      "AckMode": {
        "description": "When an input acknowledges a message to its broker — see \"acknowledgement\nmodes\" in the guide for the reasoning and, importantly, its current scope.",
        "oneOf": [
          {
            "const": "on_receipt",
            "description": "Acknowledge as soon as the message arrives, before any transform or\noutput has touched it. The default, and the behaviour every input has\nalways had — a crash between receipt and output can lose the message.",
            "type": "string"
          },
          {
            "const": "on_delivery",
            "description": "Acknowledge once the message has left *this* pipeline: every output\nthis pipeline owns has returned, successfully or not, and every\ndownstream pipeline fed from here has accepted it into its inbox. A\nfailing output does not hold up the acknowledgement — see the\narchitecture notes on why that is the current line, not a permanent\none. Not yet propagated any further than this pipeline: a downstream\npipeline's own outputs are not waited on.",
            "type": "string"
          }
        ]
      },
...
      "ApiError": {
        "description": "The error body every failing request comes back with.\n\nA Rust type rather than a hand-written schema because it has to stay in step\nwith what `AppError` actually serializes — `an_error_body_matches_the_documented_shape`\nin `tests/api.rs` is what says so.",
        "properties": {
          "error": {
            "description": "What went wrong, as one line. `anyhow`'s context chain is rendered into\nit, so the cause is in there as \"context: cause\" rather than nested.",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "DummyConfig": {
        "description": "Emits one generated message on a fixed interval — a heartbeat for testing a\npipeline without a real source attached.\n\nEvery message carries a `value` and the `current_time` it was emitted at.\nWhat the `value` holds is the `payload` field's business: a number sampled\nfrom a sine wave, so a chart of it has a shape, or a random sentence, so a\ntext transform has something to chew on.",
        "properties": {
          "amplitude": {
            "description": "peak of the sine wave — it swings between `-amplitude` and `+amplitude`.\nNumeric payloads only; defaults to 1.",
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "duration": {
            "description": "seconds between messages",
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DummyPayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "what each message's `value` holds: a `number` sampled from a sine wave,\nor a random sentence as `text`. Defaults to `number`."
          },
          "period": {
            "description": "seconds for one full turn of the sine wave. Numeric payloads only;\ndefaults to 60. Sampling is by wall clock rather than by message count,\nso the wave keeps its period whatever `duration` is.",
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          }
        },
        "required": [
          "duration"
        ],
        "title": "dummy",
        "type": "object"
      },
      "DummyPayload": {
        "description": "What a dummy input puts in each message's `value`.",
        "oneOf": [
          {
            "const": "number",
            "description": "a number sampled from a sine wave",
            "type": "string"
          },
          {
            "const": "text",
            "description": "a random sentence",
            "type": "string"
          }
        ]
      },
...
      "HttpInputConfig": {
        "description": "Accepts messages posted to this pipeline's own endpoint,\n`POST /api/pipelines/{id}/messages` — the pipeline is the receiving end of\nan http API rather than something that reaches out to a broker.\n\nThe endpoint is derived from the pipeline's id and appears as soon as the\npipeline is running; nothing is configured about it here. The body is one\nJSON message or an array of them, and an array arrives as one batch. A\npipeline can only have one of these — two would share an endpoint, and which\nof them a request went to would be a coin toss — so a second one fails to\nbuild.",
        "properties": {
          "auth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HttpAuthConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "what a post must present to be accepted. Absent — the default — means\nthe endpoint takes anything that reaches it, which is what every\npipeline with an `http` input has always done."
          },
          "capacity": {
            "description": "how many posted batches may queue up ahead of the pipeline before it\nstarts refusing them with a `503`. Defaults to 1024. The queue is what\nlets a burst through; refusing past it is deliberate, since the\nalternative is holding a request open until the pipeline catches up.",
            "format": "uint",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "title": "http",
        "type": "object"
      },
...
      "InputConfig": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/DummyConfig",
            "properties": {
              "type": {
                "const": "dummy",
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          {
            "$ref": "#/components/schemas/HttpInputConfig",
            "properties": {
              "type": {
                "const": "http",
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          {
            "$ref": "#/components/schemas/KafkaConfig",
            "properties": {
              "type": {
                "const": "kafka",
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          {
            "$ref": "#/components/schemas/NatsConfig",
            "properties": {
              "type": {
                "const": "nats",
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          {
            "$ref": "#/components/schemas/PipelineConfig",
            "properties": {
              "type": {
                "const": "pipeline",
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          {
            "$ref": "#/components/schemas/MqttConfig",
            "properties": {
              "type": {
                "const": "mqtt",
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          {
            "$ref": "#/components/schemas/RedisConfig",
            "properties": {
              "type": {
                "const": "redis",
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          {
            "$ref": "#/components/schemas/OpcuaConfig",
            "properties": {
              "type": {
                "const": "opcua",
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          }
        ],
        "properties": {
          "ack": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AckMode"
              },
              {
                "type": "null"
              }
            ],
            "description": "when this input tells its broker a message is done with. Available on\nevery input kind in the schema, but only honoured by ones with a\nbroker-side notion of \"received\" vs \"delivered\" of their own (`kafka`,\nfor now) — an input with nothing to acknowledge refuses to build rather\nthan silently treating this as `on_receipt`. Defaults to `on_receipt`,\nwhich is what every input has always done. See \"acknowledgement modes\"\nin the guide."
          },
          "buffer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BufferConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "batch messages from this input before the transforms see them — by\ncount (`static`), by time (`tumbling`) or by whichever comes first\n(`batch`). Never emits an empty batch. Available on every input kind.\nNot to be confused with the `buffer` transform."
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EnvelopeConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "attach metadata about where each message came from — the subject, topic,\npartition and so on listed under \"metadata\" below. Available on every\ninput kind. Omit it and messages are passed on exactly as they arrive."
          }
        },
        "type": "object"
      },
      "KafkaConfig": {
        "description": "Consumes JSON messages from a kafka topic, each emitted as a batch of one.\n\nA payload that isn't JSON is skipped with a warning rather than taking the\npipeline down, same as the nats input. The consumer connects on the first\nread and joins a consumer group, so kafka remembers where this pipeline got\nto between restarts.",
        "properties": {
          "connection": {
            "description": "name of the kafka connection to consume from — see \"connections\" in the\nreadme. The brokers are declared once, in the connections file, rather\nthan repeated in every pipeline reading from the same cluster.",
            "type": "string",
            "x-connection": "kafka"
          },
          "group": {
            "description": "consumer group id. Kafka tracks the read position per group, so two\npipelines sharing a group split the topic between them, and two with\ndifferent groups each get every message.",
            "type": "string"
          },
          "max_batch": {
            "description": "most messages to put in one batch. Defaults to 1 — one message per\nbatch, which is what this input has always done.\n\nRaising it only ever coalesces records that had *already arrived*: the\ninput still returns as soon as it has one, so an idle topic is no slower\nthan it was. It is worth raising when a consumer is catching up on a\nbacklog, where one-message batches make the run loop, the transforms and\nevery downstream pipeline do their per-batch work a hundred times over.",
            "format": "uint",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "start_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KafkaStartAt"
              },
              {
                "type": "null"
              }
            ],
            "description": "where to start when the group has no committed position yet: `earliest`\nreplays the topic from the beginning, `latest` only sees new messages.\nDefaults to `latest`."
          },
          "topic": {
            "description": "the topic to consume from",
            "type": "string"
          }
        },
        "required": [
          "connection",
          "topic",
          "group"
        ],
        "title": "kafka",
        "type": "object"
      },
...
      "MqttConfig": {
        "description": "Subscribes to an mqtt topic — or a topic *filter*, since mqtt's `+` and `#`\nwildcards are valid here. Each message is parsed as JSON and emitted as a\nbatch of one; a payload that isn't JSON is skipped with a warning rather\nthan taking the pipeline down, the same rule every other input follows.\n\nThe connection is opened on the first read, and a stable client id is\nderived from the pipeline's id and this topic — not configurable, since\nnothing about it is a choice this pipeline needs to make and getting it\nwrong (two inputs sharing one id) silently drops one of them.",
        "properties": {
          "connection": {
            "description": "name of the mqtt connection to subscribe on — see \"connections\" in the\nreadme. The broker it points at is declared once, in the connections\nfile, rather than repeated in every pipeline that uses it.",
            "type": "string",
            "x-connection": "mqtt"
          },
          "max_batch": {
            "description": "most messages to put in one batch. Defaults to 1 — one message per\nbatch, which is what this input has always done.\n\nRaising it only ever coalesces messages that had *already arrived*: the\ninput still returns as soon as it has one, so a quiet topic is no\nslower than it was.",
            "format": "uint",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "qos": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MqttQos"
              },
              {
                "type": "null"
              }
            ],
            "description": "the quality of service to subscribe with. Defaults to `at_most_once`.\n`ack: on_delivery` needs at least `at_least_once` here — a QoS-0\nsubscription has nothing for it to acknowledge."
          },
          "topic": {
            "description": "the topic, or topic filter, to subscribe to",
            "type": "string"
          }
        },
        "required": [
          "connection",
          "topic"
        ],
        "title": "mqtt",
        "type": "object"
      },
...
      "NatsConfig": {
        "description": "Subscribes to a nats subject. Each message is parsed as JSON and emitted as\na batch of one; a payload that isn't JSON is skipped with a warning rather\nthan taking the pipeline down. The connection is opened on the first read.",
        "properties": {
          "connection": {
            "description": "name of the nats connection to subscribe on — see \"connections\" in the\nreadme. The server it points at is declared once, in the connections\nfile, rather than repeated in every pipeline that uses it.",
            "type": "string",
            "x-connection": "nats"
          },
          "max_batch": {
            "description": "most messages to put in one batch. Defaults to 1 — one message per\nbatch, which is what this input has always done.\n\nRaising it only ever coalesces messages that had *already arrived*: the\ninput still returns as soon as it has one, so a quiet subject is no\nslower than it was.",
            "format": "uint",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "subject": {
            "description": "the subject to subscribe to",
            "type": "string"
          }
        },
        "required": [
          "connection",
          "subject"
        ],
        "title": "nats",
        "type": "object"
      },
...
      "PipelineConfig": {
        "description": "Takes another pipeline's output as its input. This is what makes the\npipelines a graph: several pipelines can read from the same upstream, and it\nfans out to all of them. The upstream must already exist when this pipeline\nis created, so declare it earlier in the config file.",
        "properties": {
          "upstream": {
            "description": "id of the pipeline to read from",
            "type": "string",
            "x-pipeline-id": true
          }
        },
        "required": [
          "upstream"
        ],
        "title": "pipeline",
        "type": "object"
      },
...
      "RedisConfig": {
        "description": "Subscribes to a redis channel. Each message is parsed as JSON and emitted\nas a batch of one; a payload that isn't JSON is skipped with a warning\nrather than taking the pipeline down. The connection is opened on the\nfirst read.\n\nPlain `SUBSCRIBE`, not `PSUBSCRIBE` — a channel name is exact, the same\nchoice the nats input makes for a subject with no wildcard. Redis pub/sub\nhas no broker-side redelivery of any kind: an unsubscribed client simply\nmisses whatever was published while it was gone, and there is nothing an\nack could hold open — the same limitation `NatsConfig` has, for the same\nreason.",
        "properties": {
          "channel": {
            "description": "the channel to subscribe to",
            "type": "string"
          },
          "connection": {
            "description": "name of the redis connection to subscribe on — see \"connections\" in\nthe readme. The server it points at is declared once, in the\nconnections file, rather than repeated in every pipeline that uses it.",
            "type": "string",
            "x-connection": "redis"
          },
          "max_batch": {
            "description": "most messages to put in one batch. Defaults to 1 — one message per\nbatch, which is what this input has always done.\n\nRaising it only ever coalesces messages that had *already arrived*: the\ninput still returns as soon as it has one, so a quiet channel is no\nslower than it was.",
            "format": "uint",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "connection",
          "channel"
        ],
        "title": "redis",
        "type": "object"
      },
...
      "SampleRequest": {
        "description": "Take a few messages from an input, without creating a pipeline.",
        "properties": {
          "input": {
            "$ref": "#/components/schemas/InputConfig",
            "description": "the input to read from, exactly as it would be written in a pipeline —\nincluding its `envelope`, so the metadata fields the messages will\nreally carry are in the sample too."
          },
          "max_messages": {
            "description": "how many messages to take. Defaults to [`DEFAULT_MAX_MESSAGES`], capped\nat [`MAX_MESSAGES`].",
            "format": "uint",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "timeout_ms": {
            "description": "how long to wait for them, in milliseconds. Defaults to\n[`DEFAULT_TIMEOUT_MS`], capped at [`MAX_TIMEOUT_MS`]. The wait ends as\nsoon as enough messages have arrived.",
            "format": "uint64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "input"
        ],
        "title": "sample request",
        "type": "object"
      },
      "SampleResponse": {
        "description": "What a sample found.\n\nNote there is no separate \"nothing arrived\" arm: an empty `messages` is\nexactly that, and it is an ordinary answer rather than a failure. A stream\nnobody is publishing to is a real state of the world, and one the user\noften needs to be shown rather than protected from.",
        "oneOf": [
          {
            "properties": {
              "messages": {
                "description": "the messages, exactly as the pipeline's first transform would have\nseen them.",
                "items": true,
                "type": "array"
              },
              "notes": {
                "description": "what the sample did differently from the pipeline it is standing in\nfor — a throwaway consumer group, an ignored buffer. Empty when it\ndid nothing differently.",
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              "outcome": {
                "const": "sampled",
                "type": "string"
              },
              "waited_ms": {
                "description": "how long it waited, in milliseconds.",
                "format": "uint64",
                "minimum": 0,
                "type": "integer"
              }
            },
            "required": [
              "outcome",
              "messages",
              "waited_ms"
            ],
            "type": "object"
          },
          {
            "properties": {
              "message": {
                "type": "string"
              },
              "outcome": {
                "const": "failed",
                "type": "string"
              },
              "stage": {
                "$ref": "#/components/schemas/SampleStage"
              }
            },
            "required": [
              "outcome",
              "stage",
              "message"
            ],
            "type": "object"
          }
        ],
        "title": "sample response"
      },
      "SampleStage": {
        "description": "Where a sample went wrong, when it did.",
        "oneOf": [
          {
            "const": "build",
            "description": "The input could not be built at all — a connection that isn't\nconfigured, a secret that doesn't resolve. The same failure creating\nthe pipeline would have given, which is the point of building the real\nthing.",
            "type": "string"
          },
          {
            "const": "read",
            "description": "It was built, and reading from it failed — the broker refused the\nsubscription, the host is unreachable.",
            "type": "string"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "description": "HTTP Basic credentials, checked against the accounts in the server's `--server-config` file. This is what anything that is not a browser should use.",
        "scheme": "basic",
        "type": "http"
      },
      "cookieAuth": {
        "description": "A session cookie from `POST /api/auth/login`. It exists for browsers: `EventSource`, which `GET /events` is consumed with, cannot send headers, so Basic credentials are not available on the one endpoint the UI needs most.",
        "in": "cookie",
        "name": "kayak_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Graph-based stream processing: an HTTP API over configurable `input → transforms → output` pipelines.\n\nEdits through this API apply to the running graph immediately and write nothing to disk — the config file is a load source and a save target, never a mirror of the runtime. `POST /api/config/save` is the only thing that writes it, and `POST /api/config/revert` is the only undo.",
    "title": "kayak",
    "version": "0.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/inputs/sample": {
      "post": {
        "description": "Configuring a stream you cannot see is guesswork, and every field reference downstream — a column's `field`, a filter's comparison — is a name someone had to already know. This builds the input in the body exactly as a pipeline would, takes up to `max_messages` from it within `timeout_ms`, and drops it.\n\nThe **real** input, including its `envelope`, so the metadata fields the messages will actually carry are in the sample too. Its `buffer` is the one thing ignored: a buffer's job is to make the pipeline wait, which is not what a sample is for. Anything the sample did differently comes back in `notes`.\n\n**Sampling is not free for every kind of input, and the ones where it isn't say so.** A kafka sample runs under a throwaway consumer group, so it neither rebalances the pipeline's group nor commits on its behalf; an mqtt sample connects under a client id of its own, because a broker disconnects the older client holding one. An `http` input is refused outright with a 400 — it is posted to rather than read from, so there is nothing to fetch.\n\n**No messages is a 200 with an empty list.** A subject nobody is publishing to is a real state of the world and the answer to the question asked; none of these inputs can replay what was published before the sample started.",
        "operationId": "sampleInput",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleRequest"
              }
            }
          },
          "description": "The input to read from, and how much to take.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleResponse"
                }
              }
            },
            "description": "The sample was taken — `outcome` says whether it produced messages or failed on the way."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The request itself was wrong: malformed JSON, an input that isn't a kind of input, or one that cannot be sampled at all."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No credentials, or credentials that were not recognised."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Signed in, but without the `admin` role this endpoint needs."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Something went wrong on the server. The body says what."
          }
        },
        "security": [
          {
            "basicAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "Fetch a few real messages from an input, without creating a pipeline",
        "tags": [
          "pipelines"
        ],
        "x-kayak-access": "admin"
      }
    },
...
  },
...
}